AuctipusItemButton.lua: Fix nil error when an auction expires.
When an auction expires, the auction.quality value goes to -1 for a
short period of time before the auction is removed.  The
ITEM_QUALITY_COLORS list doesn't have a -1 index so then we end up
trying to set the color of the auction entry to nil which blows up.
Don't set the color if a color isn't available.