AuctipusBrowseFrame.lua: Deal with synchronous callback properly.
PlaceAuctionBid(), deep down in the guys of ASearcher, could trigger a
synchronous UI_ERROR_MESSAGE event which we would handle and transition
on.  However, the transition would then happen before we had set
auction.missing to true, so the new state would be inconsistent with
what we expect.  Ensure that we set the full state in the browse frame
before we try to buy an auction so that we'll handle any immediate
errors properly.