Auctipus: Add support for shift-clicking from the craft UI.
The tradeskill UI (such as tailoring) is different from the craft UI
(such as enchanting), even though they look mostly the same.  The craft
UI doesn't have the "create all" button or edit field to select the
number of items to craft at once.  It also isn't as easily-hookable as
the tradeskill UI; we hook the UI's reagent button frames directly here.