AuctipusScan.lua: Handle nil auctions better.
The full scan seems to return a whole bunch of nil auctions on the 1.15
client.  These nil auctions do eventually populate if we wait long
enough, but re-querying all of the non-nil auctions seems to prevent
that from happening now.  So instead of processing the full page every
time we get a list update, just process the missing auctions.  The
auction indices shouldn't change.