AuctipusAuctionsFrame.lua: Generate an item link from the locked inventory item for the current auction; use it to estimate price history.
Blizzard doesn't give us a direct way to get an item link for the
auction we are currently creating.  This is needed to be able to decode
an item's suffix ID ("of the Bear").  A suffix ID is needed to
accurately get price history from the history database since all of the
different suffixes of course go for different prices.

Auctipus now checks the current list of auctions and auto-fills the
pricing based on the following criteria:

1. If the cheapest auction is your own (maybe you filed 5 stacks of
   something and then had a couple things leftover), match it.
2. If the item is up for auction by someone else, undercut it by 1
   copper.
3. If the item is not up for auction at all, set it to the last-seen
   lowest price.
4. If the item is no up for auction and no history exists, leave the
   pricing blank.